+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Dark content auto bg color and overrides - [#195](https://github.com/hkdb/aerion/issues/195)
 - Added guard rails to prevent accidental close of dialogs - [#201](https://github.com/hkdb/aerion/issues/201) - [#198](https://github.com/hkdb/aerion/issues/198)
 - Fixed message list on folder switch bug - [#200](https://github.com/hkdb/aerion/issues/200)
+- Fixed detached composer cancel orphaned draft edge case
 - Code cleanup prior to diving into v0.3.0